Re-use code between write-post and edit-post
