+++ app/write-post/page.tsx
@@ -1,38 +1,14 @@
 "use client";
 
-import { useState } from "react";
 import { useRouter } from "next/navigation";
 import { createPostAction } from "@/app/actions/post";
+import { PostForm } from "@/components/posts/PostForm";
 
 export default function WritePostPage() {
   const router = useRouter();
-  const [title, setTitle] = useState("");
-  const [content, setContent] = useState("");
-  const [imageFile, setImageFile] = useState<File | null>(null);
-  const [previewUrl, setPreviewUrl] = useState<string | null>(null);
-  const [error, setError] = useState<string | null>(null);
-  const [isLoading, setIsLoading] = useState(false);
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError(null);
-    setIsLoading(true);
-
-    const formData = new FormData();
-    formData.append("title", title);
-    formData.append("content", content);
-    if (imageFile) {
-      formData.append("image", imageFile);
-    }
-
-    const result = await createPostAction(formData);
-
-    if (!result.success) {
-      setError(result.error);
-      setIsLoading(false);
-    } else {
-      router.push(`/posts/${result.data.post._id}`);
-    }
+  const handleSuccess = (data: { post: { _id: string } }) => {
+    router.push(`/posts/${data.post._id}`);
   };
 
   return (
@@ -43,100 +19,11 @@ export default function WritePostPage() {
         </h1>
         <p className="text-gray-600">Share your thoughts with the community</p>
       </div>
-
-      {error && (
-        <div className="mb-6 p-4 bg-red-50 border border-red-200 rounded-lg">
-          <p className="text-red-600">{error}</p>
-        </div>
-      )}
-
-      <form onSubmit={handleSubmit} className="space-y-6">
-        <div>
-          <label
-            htmlFor="title"
-            className="block text-sm font-medium text-gray-700 mb-2"
-          >
-            Title
-          </label>
-          <input
-            id="title"
-            type="text"
-            value={title}
-            onChange={(e) => setTitle(e.target.value)}
-            className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-            placeholder="Enter your post title..."
-            required
-            disabled={isLoading}
-          />
-        </div>
-
-        <div>
-          <label
-            htmlFor="content"
-            className="block text-sm font-medium text-gray-700 mb-2"
-          >
-            Content
-          </label>
-          <textarea
-            id="content"
-            value={content}
-            onChange={(e) => setContent(e.target.value)}
-            rows={10}
-            className="w-full px-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent resize-vertical"
-            placeholder="Write your post content..."
-            required
-            disabled={isLoading}
-          />
-        </div>
-
-        <div>
-          <label
-            htmlFor="image"
-            className="block text-sm font-medium text-gray-700 mb-2"
-          >
-            Image (optional)
-          </label>
-          <input
-            id="image"
-            type="file"
-            accept="image/*"
-            onChange={(e) => {
-              const file = e.target.files?.[0] || null;
-              setImageFile(file);
-              setPreviewUrl(file ? URL.createObjectURL(file) : null);
-            }}
-            disabled={isLoading}
-            className="block w-full text-sm text-gray-700 border border-dashed border-gray-300 rounded-lg px-4 py-3 bg-white cursor-pointer hover:border-blue-400"
-          />
-          {previewUrl && (
-            <div className="mt-3">
-              <p className="text-sm text-gray-600 mb-2">Preview:</p>
-              <img
-                src={previewUrl}
-                alt="Preview"
-                className="w-full max-w-md rounded-lg border object-cover"
-              />
-            </div>
-          )}
-        </div>
-
-        <div className="flex gap-4">
-          <button
-            type="submit"
-            disabled={isLoading}
-            className="px-6 py-2 bg-blue-600 text-white font-medium rounded-lg hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            {isLoading ? "Creating..." : "Create Post"}
-          </button>
-          <button
-            type="button"
-            onClick={() => router.push("/")}
-            className="px-6 py-2 bg-gray-200 text-gray-800 font-medium rounded-lg hover:bg-gray-300 focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2"
-          >
-            Cancel
-          </button>
-        </div>
-      </form>
+      <PostForm
+        action={createPostAction}
+        buttonText="Create Post"
+        onSuccess={handleSuccess}
+      />
     </div>
   );
 }
